fix(server,eval): keep isAcpProvider dependency-light + await createLanguageModel in eval orchestrator
Two CI breakages from the unification commits, both invisible locally until the cross-file test isolation kicked in.

chat-service.test.ts started failing with "Export named 'simulateReadableStream' not found in module 'ai/dist/index.mjs'" because the test mocks the 'ai' module to expose only createAgentUIStreamResponse. The earlier provider-types commit added `import { isAcpProvider } from '../../agent/provider-factory'` to chat-service.ts, and provider-factory.ts transitively imports lib/clients/llm/mock-language-model.ts which does `import { simulateReadableStream } from 'ai'`. The mock kicks in before that transitive load, simulateReadableStream is undefined on the mocked module, the test file errors at import time.

Extracts the ACP-detection helpers into a new dependency-light file at apps/server/src/agent/acp-providers.ts holding only ACP_PROVIDER_TYPES + isAcpProvider, and switches chat-service.ts (and ai-sdk-agent.ts) to import from there. provider-factory.ts now consumes the same helpers and re-exports isAcpProvider so existing external callers keep working. The new file's only dep is LLM_PROVIDERS from the shared schema, so loading it does NOT pull in the ai-sdk factory graph, and chat-service.test.ts can mock 'ai' safely.

The runner / Typecheck job was failing because the earlier process-leak fix changed createLanguageModel's return shape from LanguageModel to { model, close? }, and the eval orchestrator at apps/eval/src/agents/orchestrator-executor/orchestrator-agent.ts was still doing `const model = createLanguageModel(resolvedConfig)`. Makes OrchestratorAgent.create async, destructures `{ model } = await createLanguageModel(...)`, and threads the await through the single caller in apps/eval/src/agents/orchestrator-executor/index.ts.

Verified: apps/server/tests/api/services/chat-service.test.ts now passes (was failing on the simulateReadableStream import), the full server-api suite passes 106/106, and typecheck is clean across apps/server, apps/agent, and apps/eval.
