+++ packages/browseros-agent/apps/server/src/agent/ai-sdk-agent.ts
@@ -25,6 +25,7 @@ import { metrics } from '../lib/metrics'
 import { buildFilesystemToolSet } from '../tools/filesystem/build-toolset'
 import type { ToolContext } from '../tools/framework'
 import type { ToolRegistry } from '../tools/tool-registry'
+import { isAcpProvider } from './acp-providers'
 import { CHAT_MODE_ALLOWED_TOOLS } from './chat-mode'
 import { createCompactionPrepareStep, type StepWithUsage } from './compaction'
 import { buildMcpServerSpecs, createMcpClients } from './mcp-builder'
@@ -33,7 +34,7 @@ import {
   normalizeMessagesForModel,
 } from './message-normalization'
 import { buildSystemPrompt } from './prompt'
-import { createLanguageModel, isAcpProvider } from './provider-factory'
+import { createLanguageModel } from './provider-factory'
 import { readSoulPrompt } from './soul-prompt'
 import { buildBrowserToolSet } from './tool-adapter'
 import type { ResolvedAgentConfig } from './types'
